refactor: consolidate formatter definitions (Phase 04)
Phase 04A - formatDuration:
- Added 6 new canonical variants to shared/formatters.ts: formatDurationHuman,
  formatDurationCompact, formatDurationVerbose, formatDurationParts,
  formatDurationDecimal, estimateTokensFromLogs, formatTimestamp
- Replaced 20 local formatDuration definitions across UsageDashboard (10),
  AboutModal, FirstRunCelebration, Toast, CueModal, CuePanel, SymphonyModal,
  useContributorStats, cli/formatter, tabExport, groupChatExport
- Updated canonical formatNumber to use uppercase K and toString() for <1000

Phase 04B - formatElapsedTime:
- Replaced 5 local formatElapsedTime definitions in MergeProgressModal,
  MergeProgressOverlay, SummarizeProgressModal, SummarizeProgressOverlay,
  TransferProgressModal with import from shared/formatters

Phase 04C - formatTimestamp:
- Added canonical formatTimestamp(timestamp, style) with 4 styles:
  time, datetime, smart (default), full
- Replaced 13 local formatTime/formatTimestamp definitions across
  HistoryEntryItem, GroupChatHistoryPanel, MobileHistoryPanel, MessageHistory,
  GroupChatPanel (smart); WizardMessageBubble, ConversationScreen,
  LongestAutoRunsTable (time); HistoryDetailModal, ResponseViewer (datetime);
  tabExport, groupChatExport (full)
- GroupChatMessages and ParticipantCard refactored to delegate string
  formatting to shared formatTimestamp
- ThinkingStatusPill skipped (duration formatter, not timestamp)

Phase 04D - formatNumber, formatFileSize, estimateTokens, stripAnsi, generateId:
- Replaced 5 local formatNumber in UsageDashboard components
- Replaced 2 formatFileSize (documentStats, filePreviewUtils) with
  thin wrappers around shared formatSize
- Replaced 4 estimateTokens/estimateTokensFromLogs (MergeSessionModal,
  SendToAgentModal, useSendToAgent, useMergeSession); tokenCounter.ts
  now delegates to shared estimateTokenCount
- Replaced main/utils/stripAnsi.ts with re-export from shared/stringUtils
- Replaced main/stats/utils.ts generateId with re-export from shared/uuid
- Skipped symphony.ts formatNumber (different function: toLocaleString)
- Skipped useLayerStack/useBatchedSessionUpdates/useCommandHistory/
  useOfflineQueue generateId (domain-specific ID formats with prefixes)
- Skipped contextExtractor.ts estimateTokenCount (different signature)
