+++ src/__tests__/renderer/components/FilePreview/filePreviewUtils.test.ts
@@ -138,16 +138,16 @@ describe('filePreviewUtils', () => {
 		});
 
 		it('formats kilobytes', () => {
-			expect(formatFileSize(1024)).toBe('1 KB');
+			expect(formatFileSize(1024)).toBe('1.0 KB');
 			expect(formatFileSize(1536)).toBe('1.5 KB');
 		});
 
 		it('formats megabytes', () => {
-			expect(formatFileSize(1048576)).toBe('1 MB');
+			expect(formatFileSize(1048576)).toBe('1.0 MB');
 		});
 
 		it('formats gigabytes', () => {
-			expect(formatFileSize(1073741824)).toBe('1 GB');
+			expect(formatFileSize(1073741824)).toBe('1.0 GB');
 		});
 	});
