+++ src/__tests__/renderer/components/TransferProgressModal.test.tsx
@@ -430,7 +430,7 @@ describe('TransferProgressModal', () => {
 			);
 
 			expect(screen.getByText('Elapsed:')).toBeInTheDocument();
-			expect(screen.getByText('0s')).toBeInTheDocument();
+			expect(screen.getByText('0ms')).toBeInTheDocument();
 		});
 
 		it('updates elapsed time every second', async () => {
@@ -445,7 +445,7 @@ describe('TransferProgressModal', () => {
 				/>
 			);
 
-			expect(screen.getByText('0s')).toBeInTheDocument();
+			expect(screen.getByText('0ms')).toBeInTheDocument();
 
 			// Use act to wrap the timer advancement
 			await act(async () => {
