+++ src/renderer/components/AboutModal.tsx
@@ -16,6 +16,7 @@ import type { GlobalAgentStats } from '../../shared/types';
 import { MODAL_PRIORITIES } from '../constants/modalPriorities';
 import { AchievementCard } from './AchievementCard';
 import { formatTokensCompact } from '../utils/formatters';
+import { formatDurationHuman } from '../../shared/formatters';
 import { Modal } from './ui/Modal';
 import { buildMaestroUrl } from '../utils/buildMaestroUrl';
 import { openUrl } from '../utils/openUrl';
@@ -89,20 +90,7 @@ export function AboutModal({
 
 	// formatTokensCompact and formatSize imported from ../utils/formatters
 
-	// Format duration from milliseconds
-	const formatDuration = (ms: number): string => {
-		const seconds = Math.floor(ms / 1000);
-		const minutes = Math.floor(seconds / 60);
-		const hours = Math.floor(minutes / 60);
-
-		if (hours > 0) {
-			return `${hours}h ${minutes % 60}m`;
-		}
-		if (minutes > 0) {
-			return `${minutes}m ${seconds % 60}s`;
-		}
-		return `${seconds}s`;
-	};
+	const formatDuration = formatDurationHuman;
 
 	// Custom escape handler that checks for badge overlay first
 	// Uses refs to avoid dependency changes that would cause infinite loops
