+++ src/renderer/components/InlineWizard/WizardMessageBubble.tsx
@@ -22,6 +22,7 @@ import {
 	REMARK_GFM_PLUGINS,
 	createWizardBubbleMarkdownComponents,
 } from '../../utils/markdownConfig';
+import { formatTimestamp } from '../../../shared/formatters';
 
 /**
  * Message structure for wizard conversations
@@ -56,13 +57,6 @@ export interface WizardMessageBubbleProps {
 	) => void;
 }
 
-/**
- * Format timestamp for display
- */
-function formatTimestamp(timestamp: number): string {
-	const date = new Date(timestamp);
-	return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-}
 
 /**
  * WizardMessageBubble - Individual conversation message display for inline wizard
@@ -184,7 +178,7 @@ export const WizardMessageBubble = React.memo(function WizardMessageBubble({
 					}}
 					data-testid="message-timestamp"
 				>
-					{formatTimestamp(message.timestamp)}
+					{formatTimestamp(message.timestamp, 'time')}
 				</div>
 			</div>
 		</div>
