+++ src/renderer/components/SymphonyModal.tsx
@@ -64,6 +64,7 @@ import {
 import { formatShortcutKeys } from '../utils/shortcutFormatter';
 import { buildMaestroUrl } from '../utils/buildMaestroUrl';
 import { openUrl } from '../utils/openUrl';
+import { formatDurationCompact as formatDurationMs } from '../../shared/formatters';
 
 // ============================================================================
 // Types
@@ -126,12 +127,7 @@ function formatCacheAge(cacheAgeMs: number | null): string {
 	return 'just now';
 }
 
-function formatDurationMs(ms: number): string {
-	const totalSeconds = Math.floor(ms / 1000);
-	if (totalSeconds < 60) return `${totalSeconds}s`;
-	if (totalSeconds < 3600) return `${Math.floor(totalSeconds / 60)}m`;
-	return `${Math.floor(totalSeconds / 3600)}h ${Math.floor((totalSeconds % 3600) / 60)}m`;
-}
+
 
 function formatDate(isoString: string): string {
 	return new Date(isoString).toLocaleDateString('en-US', {
