+++ src/renderer/components/UsageDashboard/AutoRunStats.tsx
@@ -17,6 +17,7 @@ import { Play, CheckSquare, ListChecks, Target, Clock, Timer } from 'lucide-reac
 import type { Theme } from '../../types';
 import type { StatsTimeRange } from '../../hooks/stats/useStats';
 import { captureException } from '../../utils/sentry';
+import { formatDurationHuman as formatDuration, formatNumber } from '../../../shared/formatters';
 
 /**
  * Auto Run session data shape from the API
@@ -42,40 +43,7 @@ interface AutoRunStatsProps {
 	columns?: number;
 }
 
-/**
- * Format duration in milliseconds to human-readable string
- * Examples: "12h 34m", "5m 30s", "45s"
- */
-function formatDuration(ms: number): string {
-	if (ms === 0) return '0s';
 
-	const totalSeconds = Math.floor(ms / 1000);
-	const hours = Math.floor(totalSeconds / 3600);
-	const minutes = Math.floor((totalSeconds % 3600) / 60);
-	const seconds = totalSeconds % 60;
-
-	if (hours > 0) {
-		return `${hours}h ${minutes}m`;
-	}
-	if (minutes > 0) {
-		return `${minutes}m ${seconds}s`;
-	}
-	return `${seconds}s`;
-}
-
-/**
- * Format large numbers with K/M suffixes for readability
- * Examples: "1.2K", "3.5M", "42"
- */
-function formatNumber(num: number): string {
-	if (num >= 1000000) {
-		return `${(num / 1000000).toFixed(1)}M`;
-	}
-	if (num >= 1000) {
-		return `${(num / 1000).toFixed(1)}K`;
-	}
-	return num.toString();
-}
 
 /**
  * Single metric card component
