+++ src/renderer/components/UsageDashboard/LocationDistributionChart.tsx
@@ -16,6 +16,7 @@ import { memo, useState, useMemo } from 'react';
 import type { Theme } from '../../types';
 import type { StatsAggregation } from '../../hooks/stats/useStats';
 import { COLORBLIND_BINARY_PALETTE } from '../../constants/colorblindPalettes';
+import { formatNumber } from '../../../shared/formatters';
 
 interface LocationData {
 	location: 'local' | 'remote';
@@ -34,18 +35,7 @@ interface LocationDistributionChartProps {
 	colorBlindMode?: boolean;
 }
 
-/**
- * Format large numbers with K/M suffixes
- */
-function formatNumber(num: number): string {
-	if (num >= 1000000) {
-		return `${(num / 1000000).toFixed(1)}M`;
-	}
-	if (num >= 1000) {
-		return `${(num / 1000).toFixed(1)}K`;
-	}
-	return num.toString();
-}
+
 
 /**
  * Get a color for remote sessions that contrasts with accent (local)
