+++ src/renderer/components/UsageDashboard/LongestAutoRunsTable.tsx
@@ -18,6 +18,7 @@ import { Trophy } from 'lucide-react';
 import type { Theme } from '../../types';
 import type { StatsTimeRange } from '../../hooks/stats/useStats';
 import { captureException } from '../../utils/sentry';
+import { formatDurationHuman as formatDuration, formatTimestamp } from '../../../shared/formatters';
 
 /**
  * Auto Run session data shape from the API
@@ -43,25 +44,7 @@ interface LongestAutoRunsTableProps {
 
 const MAX_ROWS = 25;
 
-/**
- * Format duration in milliseconds to human-readable string
- */
-function formatDuration(ms: number): string {
-	if (ms === 0) return '0s';
 
-	const totalSeconds = Math.floor(ms / 1000);
-	const hours = Math.floor(totalSeconds / 3600);
-	const minutes = Math.floor((totalSeconds % 3600) / 60);
-	const seconds = totalSeconds % 60;
-
-	if (hours > 0) {
-		return `${hours}h ${minutes}m`;
-	}
-	if (minutes > 0) {
-		return `${minutes}m ${seconds}s`;
-	}
-	return `${seconds}s`;
-}
 
 /**
  * Format agent type to display name
@@ -109,15 +92,7 @@ function formatDate(timestamp: number): string {
 	});
 }
 
-/**
- * Format time for table display
- */
-function formatTime(timestamp: number): string {
-	return new Date(timestamp).toLocaleTimeString('en-US', {
-		hour: 'numeric',
-		minute: '2-digit',
-	});
-}
+const formatTime = (timestamp: number) => formatTimestamp(timestamp, 'time');
 
 export const LongestAutoRunsTable = memo(function LongestAutoRunsTable({
 	timeRange,
