+++ src/renderer/components/UsageDashboard/PeakHoursChart.tsx
@@ -15,6 +15,7 @@
 import { memo, useState, useMemo } from 'react';
 import type { Theme } from '../../types';
 import type { StatsAggregation } from '../../hooks/stats/useStats';
+import { formatDurationCompact as formatDuration } from '../../../shared/formatters';
 
 type MetricMode = 'count' | 'duration';
 
@@ -37,22 +38,7 @@ function formatHour(hour: number): string {
 	return `${hour - 12}pm`;
 }
 
-/**
- * Format duration in milliseconds to human-readable string
- */
-function formatDuration(ms: number): string {
-	const totalSeconds = Math.floor(ms / 1000);
-	const hours = Math.floor(totalSeconds / 3600);
-	const minutes = Math.floor((totalSeconds % 3600) / 60);
 
-	if (hours > 0) {
-		return `${hours}h ${minutes}m`;
-	}
-	if (minutes > 0) {
-		return `${minutes}m`;
-	}
-	return `${totalSeconds}s`;
-}
 
 export const PeakHoursChart = memo(function PeakHoursChart({
 	data,
