+++ src/renderer/components/UsageDashboard/WeekdayComparisonChart.tsx
@@ -15,6 +15,7 @@ import { memo, useMemo } from 'react';
 import { Briefcase, Coffee } from 'lucide-react';
 import type { Theme } from '../../types';
 import type { StatsAggregation } from '../../hooks/stats/useStats';
+import { formatDurationHuman as formatDuration } from '../../../shared/formatters';
 
 interface WeekdayComparisonChartProps {
 	/** Aggregated stats data from the API */
@@ -25,25 +26,7 @@ interface WeekdayComparisonChartProps {
 	colorBlindMode?: boolean;
 }
 
-/**
- * Format duration in milliseconds to human-readable string
- */
-function formatDuration(ms: number): string {
-	if (ms === 0) return '0s';
 
-	const totalSeconds = Math.floor(ms / 1000);
-	const hours = Math.floor(totalSeconds / 3600);
-	const minutes = Math.floor((totalSeconds % 3600) / 60);
-	const seconds = totalSeconds % 60;
-
-	if (hours > 0) {
-		return `${hours}h ${minutes}m`;
-	}
-	if (minutes > 0) {
-		return `${minutes}m ${seconds}s`;
-	}
-	return `${seconds}s`;
-}
 
 export const WeekdayComparisonChart = memo(function WeekdayComparisonChart({
 	data,
