+++ src/renderer/utils/tokenCounter.ts
@@ -8,6 +8,7 @@
  */
 
 import { getEncoding, type Tiktoken } from 'js-tiktoken';
+import { estimateTokenCount } from '../../shared/formatters';
 
 // Lazy-loaded tokenizer encoder (cl100k_base is used by Claude/GPT-4)
 let encoderPromise: Promise<Tiktoken> | null = null;
@@ -46,17 +47,9 @@ export async function countTokens(text: string): Promise<number> {
 
 /**
  * Synchronous token estimation using character-based heuristic.
- * Use this when you need a quick estimate without async.
- * Less accurate than countTokens() but faster and synchronous.
- *
- * @param text - The text to estimate tokens for
- * @returns Estimated token count
+ * Re-exports estimateTokenCount from shared/formatters for backwards compatibility.
  */
-export function estimateTokens(text: string): number {
-	// Average of ~4 characters per token for English text
-	// Code tends to be denser (~3 chars/token), but this is a reasonable average
-	return Math.ceil(text.length / 4);
-}
+export const estimateTokens = (text: string): number => estimateTokenCount(text);
 
 /**
  * Format a token count for display with appropriate suffix.
